+++ CHANGELOG.md
@@ -35,6 +35,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ## Unreleased
 
+### Improvements
+
+* [#227](https://github.com/babylonlabs-io/btc-staker/pull/227) feat: staker multisig support
+
 ## v0.18.1
 
 ### Bug Fixes